+++ crates/pg-sink/tests/streaming_large_txn.rs
@@ -118,7 +118,7 @@ async fn cleanup(pool: &sqlx::PgPool, admin: &tokio_postgres::Client, epoch: Epo
             let _ = store.delete(&object_store::path::Path::from(key)).await;
         }
     }
-    let _ = sqlx::query("DELETE FROM walrus.file_manifest WHERE epoch = $1")
+    let _ = sqlx::query("WITH authorized AS MATERIALIZED (SELECT set_config('walrus.manifest_delete_protocol','2',true) AS protocol) DELETE FROM walrus.file_manifest WHERE epoch = $1 AND (SELECT protocol='2' FROM authorized)")
         .bind(epoch)
         .execute(pool)
         .await;
@@ -197,16 +197,29 @@ async fn large_txn_single_ready_file_only_after_stream_commit() {
                 continue;
             };
             match &msg {
-                Message::Relation { relation, .. } => {
+                Message::Relation { relation, xid } => {
                     cache
                         .upsert_from_relation(relation.clone(), common::SchemaVersionNo(1))
                         .unwrap();
+                    if let (Some(sub_xid), Some(top_xid)) = (*xid, demux.current_top()) {
+                        demux.bind_relation(
+                            top_xid,
+                            sub_xid,
+                            relation.oid,
+                            common::SchemaVersionNo(1),
+                        );
+                    }
                 }
                 Message::StreamStart { xid, first_segment } => {
-                    demux.on_stream_start(*xid, *first_segment, frame_lsn);
-                    checkpoint.set_open_txn_floor(demux.open_floor());
+                    let pre_start_ceiling = checkpoint.capture_pre_stream_start_ceiling();
+                    demux
+                        .on_stream_start(*xid, *first_segment, frame_lsn)
+                        .unwrap();
+                    if *first_segment {
+                        checkpoint.on_stream_start(*xid, pre_start_ceiling).unwrap();
+                    }
                 }
-                Message::StreamStop => demux.on_stream_stop(),
+                Message::StreamStop => demux.on_stream_stop().unwrap(),
                 m @ (Message::Insert { xid: Some(_), .. }
                 | Message::Update { xid: Some(_), .. }
                 | Message::Delete { xid: Some(_), .. }) => {
@@ -252,7 +265,7 @@ async fn large_txn_single_ready_file_only_after_stream_commit() {
                             .await
                             .unwrap();
                     }
-                    checkpoint.set_open_txn_floor(demux.open_floor());
+                    checkpoint.on_stream_end(*xid).unwrap();
                     checkpoint.on_batch_durable(*clsn);
                     commit_lsn = Some(*clsn);
                 }
@@ -370,22 +383,35 @@ async fn whole_txn_abort_writes_no_ready_row() {
                 continue;
             };
             match &msg {
-                Message::Relation { relation, .. } => {
+                Message::Relation { relation, xid } => {
                     cache
                         .upsert_from_relation(relation.clone(), common::SchemaVersionNo(1))
                         .unwrap();
+                    if let (Some(sub_xid), Some(top_xid)) = (*xid, demux.current_top()) {
+                        demux.bind_relation(
+                            top_xid,
+                            sub_xid,
+                            relation.oid,
+                            common::SchemaVersionNo(1),
+                        );
+                    }
                 }
                 Message::StreamStart { xid, first_segment } => {
-                    demux.on_stream_start(*xid, *first_segment, frame_lsn);
+                    demux
+                        .on_stream_start(*xid, *first_segment, frame_lsn)
+                        .unwrap();
                 }
-                Message::StreamStop => demux.on_stream_stop(),
+                Message::StreamStop => demux.on_stream_stop().unwrap(),
                 m @ (Message::Insert { xid: Some(_), .. }
                 | Message::Update { xid: Some(_), .. }
                 | Message::Delete { xid: Some(_), .. }) => {
                     demux.on_change(&cache, m, &sink, frame_lsn).await.unwrap();
                 }
                 Message::StreamAbort { top_xid, sub_xid } => {
-                    demux.on_stream_abort(*top_xid, *sub_xid, &sink).await;
+                    demux
+                        .on_stream_abort(*top_xid, *sub_xid, &sink)
+                        .await
+                        .unwrap();
                     saw_abort = true;
                 }
                 // The trailing small (non-streamed) commit ends the loop.
